Clean interpreter folder
