--- src/interpreter/parser/test
@@ -1,3 +1,0 @@
-three := 3
-seven := one + two * three
-seven * seven
